Use simply express server
