+++ .gitignore
@@ -9,5 +9,3 @@ npm-debug.log
 coverage
 .DS_Store
 .nyc_output
-/debug/access-token.js
-/bench/bench.js
