+++ package.json
@@ -50,10 +50,12 @@
     "doctrine": "^1.2.1",
     "documentation": "https://github.com/documentationjs/documentation/archive/bb41619c734e59ef3fbc3648610032efcfdaaace.tar.gz",
     "documentation-theme-utils": "3.0.0",
+    "enchilada": "^0.13.0",
     "envify": "^3.4.0",
     "eslint": "^2.5.3",
     "eslint-config-mourner": "^2.0.0",
     "eslint-plugin-html": "^1.5.1",
+    "express": "^4.13.4",
     "gl": "^4.0.1",
     "handlebars": "4.0.5",
     "highlight.js": "9.3.0",
@@ -68,14 +70,12 @@
     "remark": "4.2.2",
     "remark-html": "3.0.0",
     "sinon": "^1.15.4",
-    "st": "^1.2.0",
     "tap": "^5.7.0",
     "through": "^2.3.7",
     "transform-loader": "^0.2.3",
     "unist-util-visit": "1.1.0",
     "vinyl": "1.1.1",
     "vinyl-fs": "2.4.3",
-    "watchify": "^3.7.0",
     "webpack": "^1.13.1",
     "webworkify-webpack": "^1.1.3"
   },
@@ -87,22 +87,14 @@
     "./js/util/web_worker.js": "./js/util/browser/web_worker.js"
   },
   "scripts": {
-    "build-dev": "browserify js/mapbox-gl.js --debug --standalone mapboxgl > dist/mapbox-gl-dev.js && tap --no-coverage test/build/dev.test.js",
-    "watch-dev": "watchify js/mapbox-gl.js --debug --standalone mapboxgl -o dist/mapbox-gl-dev.js -v",
-    "build-min": "browserify js/mapbox-gl.js --debug -t unassertify --plugin [minifyify --map mapbox-gl.js.map --output dist/mapbox-gl.js.map] --standalone mapboxgl > dist/mapbox-gl.js && tap --no-coverage test/build/min.test.js",
-    "watch-min": "watchify js/mapbox-gl.js --debug -t unassertify --plugin [minifyify --map mapbox-gl.js.map --output dist/mapbox-gl.js.map] --standalone mapboxgl -o dist/mapbox-gl.js -v",
-    "build-token": "browserify debug/access-token-src.js --debug -t envify > debug/access-token.js",
-    "watch-bench": "watchify bench/index.js --plugin [minifyify --no-map] -t [babelify --presets react] -t unassertify -t envify -o bench/bench.js -v",
-    "download-bench-data": "node bench/download-data.js",
-    "start-server": "st --no-cache --localhost --port 9966 --index index.html .",
-    "start": "npm run build-token && npm run download-bench-data && npm run watch-dev & npm run watch-bench & npm run start-server & wait",
-    "start-debug": "npm run build-token && npm run watch-dev & npm run start-server & wait",
-    "start-bench": "npm run build-token && npm run download-bench-data && npm run watch-bench & npm run start-server & wait",
+    "build-dev": "browserify js/mapbox-gl.js --debug --ignore-transform unassertify --standalone mapboxgl > dist/mapbox-gl-dev.js && tap --no-coverage test/build/dev.test.js",
     "build-docs": "documentation build --github --format html -c documentation.yml --theme ./docs/_theme --output docs/api/",
+    "build-min": "browserify js/mapbox-gl.js --debug --transform unassertify --plugin [minifyify --map mapbox-gl.js.map --output dist/mapbox-gl.js.map] --standalone mapboxgl > dist/mapbox-gl.js && tap --no-coverage test/build/min.test.js",
     "build": "npm run build-docs # invoked by publisher when publishing docs on the mb-pages branch",
-    "start-docs": "npm run build-min && npm run build-docs && jekyll serve -w",
-    "lint": "eslint  --ignore-path .gitignore js test bench docs/_posts/examples/*.html",
+    "lint": "eslint js test bench server.js docs/_posts/examples/*.html",
     "open-changed-examples": "git diff --name-only mb-pages HEAD -- docs/_posts/examples/*.html | awk '{print \"http://127.0.0.1:4000/mapbox-gl-js/example/\" substr($0,33,length($0)-37)}' | xargs open",
+    "start-docs": "npm run build-min && npm run build-docs && jekyll serve -w",
+    "start": "node server.js",
     "test-suite": "node test/render.test.js && node test/query.test.js",
     "test": "npm run lint && tap --reporter dot test/js/*/*.js test/build/webpack.test.js"
   }
