+++ server/bench/index.html
@@ -14,7 +14,7 @@
     <div id="benchmarks"></div>
     <script src="https://npmcdn.com/react@15.3.0/dist/react.js"></script>
     <script src="https://npmcdn.com/react-dom@15.3.0/dist/react-dom.js"></script>
-    <script src="bench.js"></script>
+    <script src="index.js"></script>
 </body>
 
 </html>
