+++ server/bench/lib/measure_framerate.js
