+++ server/bench/lib/setup_geojson_map.js
