+++ server/events.html
@@ -94,7 +94,7 @@ html, body, #map {
 
 <div id='console' class='hide'></div>
 
-<script src='../dist/mapbox-gl-dev.js'></script>
+<script src='mapbox-gl.js'></script>
 <script src='access-token.js'></script>
 
 <script>
