+++ server/image.html
@@ -14,7 +14,7 @@
 
 <div id='map'></div>
 
-<script src='../dist/mapbox-gl-dev.js'></script>
+<script src='mapbox-gl.js'></script>
 <script src='access-token.js'></script>
 
 <script>
@@ -23,7 +23,7 @@ var imageStyle = {
     "sources": {
         "image": {
             "type": "image",
-            "url": "../node_modules/mapbox-gl-test-suite/image/0.png",
+            "url": "node_modules/mapbox-gl-test-suite/image/0.png",
             "coordinates": [
                 [-122.51596391201019, 37.56238816766053],
                 [-122.51467645168304, 37.56410183312965],
