+++ server/index.html
@@ -49,13 +49,13 @@
     <em>Waiting for data...</em>
 </div>
 
-<script src='../dist/mapbox-gl-dev.js'></script>
+<script src='mapbox-gl.js'></script>
 <script src='access-token.js'></script>
 
 <script>
 var map = window.map = new mapboxgl.Map({
     container: 'map',
-    zoom: 3,
+    zoom: 12.5,
     center: [-77.01866, 38.888],
     style: 'mapbox://styles/mapbox/streets-v9',
     hash: true
@@ -65,37 +65,32 @@ map.addControl(new mapboxgl.Navigation());
 map.addControl(new mapboxgl.Geolocate());
 
 map.on('load', function() {
-
-    map.addSource('circles', {
+    map.addSource('geojson', {
         "type": "geojson",
-        "data": "../node_modules/mapbox-gl-test-suite/data/places.geojson"
+        "data": "node_modules/mapbox-gl-test-suite/data/linestring.geojson"
     });
 
     map.addLayer({
-        "id": "circles",
-        "type": "circle",
-        "source": "circles",
+        "id": "route",
+        "type": "line",
+        "source": "geojson",
         "paint": {
-            "circle-radius": {
-                property: "mapbox",
-                stops: [
-                    [{ zoom: 0, value: 0 }, 2],
-                    [{ zoom: 0, value: 100 }, 10],
-                    [{ zoom: 6, value: 0 }, 20],
-                    [{ zoom: 6, value: 100 }, 100]
-                ]
-            },
-            "circle-color": {
-                property: "mapbox",
-                stops: [
-                    [{ zoom: 0, value: 0 }, 'red'],
-                    [{ zoom: 0, value: 100 }, 'violet'],
-                    [{ zoom: 6, value: 0 }, 'blue'],
-                    [{ zoom: 6, value: 100 }, 'green']
+            "line-color": "#EC8D8D",
+            "line-width": {
+                "base": 1.5,
+                "stops": [
+                    [
+                        5,
+                        0.75
+                    ],
+                    [
+                        18,
+                        32
+                    ]
                 ]
             }
         }
-    });
+    }, 'country-label-lg');
 
     var bufferTimes = {};
     map.on('tile.stats', function(bufferTimes) {
@@ -149,7 +144,6 @@ document.onkeypress = function(e) {
         return false;
     }
 };
-
 </script>
 
 </body>
