+++ server/mapbox-gl.js
@@ -1,0 +1,1 @@
+window.mapboxgl = require('../js/mapbox-gl');
