+++ server/multiple.html
@@ -8,17 +8,18 @@
     <link rel='stylesheet' href='../dist/mapbox-gl.css' />
     <style>
         body { margin: 0; padding: 0; }
-        html, body, #map { height: 100%; }
+        .map { width: 300px; height: 300px; display: block; float: left; margin: 5px; }
     </style>
 </head>
 
-<div id='map'></div>
+<div class='map' id='map1'></div>
+<div class='map' id='map2'></div>
 
-<script src='../dist/mapbox-gl-dev.js'></script>
+<script src='mapbox-gl.js'></script>
 <script src='access-token.js'></script>
 
 <script>
-var videoStyle = {
+var style = {
     "version": 8,
     "sources": {
         "satellite": {
@@ -26,16 +27,6 @@ var videoStyle = {
             "url": "mapbox://mapbox.satellite",
             "tileSize": 256
         },
-        "video": {
-            "type": "video",
-            "urls": [ "https://www.mapbox.com/drone/video/drone.mp4", "https://www.mapbox.com/drone/video/drone.webm" ],
-            "coordinates": [
-                [-122.51596391201019, 37.56238816766053],
-                [-122.51467645168304, 37.56410183312965],
-                [-122.51309394836426, 37.563391708549425],
-                [-122.51423120498657, 37.56161849366671]
-            ]
-        }
     },
     "layers": [{
         "id": "background",
@@ -47,22 +38,40 @@ var videoStyle = {
         "id": "satellite",
         "type": "raster",
         "source": "satellite"
-    }, {
-        "id": "video",
-        "type": "raster",
-        "source": "video"
     }]
 };
 
-var map = new mapboxgl.Map({
-  container: 'map',
+var map1 = new mapboxgl.Map({
+  container: 'map1',
   minZoom: 14,
   zoom: 17,
   center: [-122.514426, 37.562984],
   bearing: -96,
-  style: videoStyle,
+  style: style,
   hash: false
 });
+
+var map2 = new mapboxgl.Map({
+  container: 'map2',
+  minZoom: 14,
+  zoom: 17,
+  center: [-122.514426, 37.562984],
+  bearing: -96,
+  style: style,
+  hash: false
+});
+
+function popup(e) {
+    if (e.originalEvent.shiftKey) return;
+    (new mapboxgl.Popup())
+        .setLngLat(e.lngLat)
+        .setHTML('<h3>' + e.point.x + ', ' + e.point.y + '</h3>')
+        .addTo(this);
+}
+
+map1.on('click', popup.bind(map1));
+map2.on('click', popup.bind(map2));
+
 </script>
 
 </body>
