+++ server/query-features.html
@@ -14,7 +14,7 @@
 
 <body>
 <div id='map'></div>
-<script src='../dist/mapbox-gl-dev.js'></script>
+<script src='mapbox-gl.js'></script>
 <script src='access-token.js'></script>
 
 <script>
