+++ server/wms.html
@@ -32,7 +32,7 @@ html, body, #map { height: 100%; }
   <input id='show-symbol-collision-boxes-checkbox' name='show-symbol-collision-boxes' type='checkbox'> <label for='show-symbol-collision-boxes'>collision debug</label><br/>
 </div>
 
-<script src='../dist/mapbox-gl-dev.js'></script>
+<script src='mapbox-gl.js'></script>
 <script src='access-token.js'></script>
 
 <script>
